requirements.txt
requests==2.26.0
sanic==20.12.4
pyjwt==2.3.0
